+++ R/check.R
@@ -169,7 +169,7 @@ check_hugo = function() {
   check_progress('Checking .Rprofile for Hugo version used by blogdown...')
 
   # .Rprofile exists + most recent Hugo
-  if (!(is.null(sv <- getOption('blogdown.hugo.version')))) {
+  if (!(is.null(sv <- get_option('blogdown.hugo.version')))) {
     check_success(sprintf('blogdown is using Hugo %s to build site locally.', sv))
   } else {
     check_progress('Hugo version not set in .Rprofile.')